+++ src/main/java/com/carbonldp/ldp/containers/ContainerRepository.java
@@ -1,6 +1,7 @@
 package com.carbonldp.ldp.containers;
 
 import com.carbonldp.descriptions.APIPreferences.ContainerRetrievalPreference;
+import com.carbonldp.http.OrderByRetrievalPreferences;
 import com.carbonldp.ldp.containers.ContainerDescription.Type;
 import com.carbonldp.ldp.sources.RDFSource;
 import org.openrdf.model.IRI;
@@ -65,4 +66,12 @@ public interface ContainerRepository {
 	public void removeMembers( IRI targetIRI, Type containerType );
 
 	public TypedContainerRepository getTypedRepository( Type containerType );
+
+	public Set<IRI> getContainmentIRIs( IRI targetIRI, OrderByRetrievalPreferences orderByRetrievalPreferences );
+
+	public Set<IRI> getContainmentIRIs( IRI targetIRI );
+
+	public Set<IRI> getMemberIRIs( IRI targetIRI, OrderByRetrievalPreferences orderByRetrievalPreferences );
+
+	public Set<IRI> getMemberIRIs( IRI targetIRI );
 }