+++ src/main/java/com/carbonldp/ldp/containers/SesameContainerService.java
@@ -5,17 +5,24 @@ import com.carbonldp.exceptions.InvalidRDFTypeException;
 import com.carbonldp.exceptions.InvalidResourceException;
 import com.carbonldp.exceptions.ResourceAlreadyExistsException;
 import com.carbonldp.exceptions.ResourceDoesntExistException;
+import com.carbonldp.http.OrderByRetrievalPreferences;
 import com.carbonldp.ldp.AbstractSesameLDPService;
+import com.carbonldp.ldp.sources.RDFSource;
+import com.carbonldp.ldp.sources.RDFSourceDescription;
 import com.carbonldp.ldp.sources.RDFSourceService;
 import com.carbonldp.models.Infraction;
+import com.carbonldp.rdf.RDFBlankNode;
 import com.carbonldp.rdf.RDFDocumentFactory;
 import com.carbonldp.rdf.RDFResource;
 import com.carbonldp.rdf.RDFResourceRepository;
+import com.carbonldp.rdf.RDFResourceDescription;
 import com.carbonldp.spring.ServicesInvoker;
-import com.carbonldp.web.exceptions.NotImplementedException;
 import org.joda.time.DateTime;
+import org.openrdf.model.BNode;
 import org.openrdf.model.IRI;
 import org.openrdf.model.Statement;
+import org.openrdf.model.ValueFactory;
+import org.openrdf.model.impl.SimpleValueFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.io.File;
@@ -29,7 +36,7 @@ public class SesameContainerService extends AbstractSesameLDPService implements 
 	private RDFResourceRepository resourceRepository;
 
 	@Override
-	public Container get( IRI containerIRI, Set<APIPreferences.ContainerRetrievalPreference> containerRetrievalPreferences ) {
+	public Container get( IRI containerIRI, Set<APIPreferences.ContainerRetrievalPreference> containerRetrievalPreferences, OrderByRetrievalPreferences orderByRetrievalPreferences ) {
 		ContainerDescription.Type containerType = getContainerType( containerIRI );
 		if ( containerType == null ) throw new InvalidRDFTypeException( ContainerDescription.Resource.CLASS.getIRI().stringValue() );
 
@@ -43,7 +50,13 @@ public class SesameContainerService extends AbstractSesameLDPService implements 
 					container.getBaseModel().addAll( containerRepository.getContainmentTriples( containerIRI ) );
 					break;
 				case CONTAINED_RESOURCES:
-					throw new NotImplementedException();
+					Set<IRI> children = containerRepository.getContainmentIRIs( containerIRI, orderByRetrievalPreferences );
+					if ( containerRetrievalPreferences.contains( APIPreferences.ContainerRetrievalPreference.MEMBER_RESOURCES ) ) {
+						Set<IRI> members = containerRepository.getMemberIRIs( containerIRI, orderByRetrievalPreferences );
+						children.addAll( members );
+					}
+					container = getResources( children, container );
+					break;
 				case MEMBERSHIP_TRIPLES:
 					if ( containerRetrievalPreferences.contains( APIPreferences.ContainerRetrievalPreference.NON_READABLE_MEMBERSHIP_RESOURCE_TRIPLES ) ) {
 						container.getBaseModel().addAll( getMembershipTriples( containerIRI ) );
@@ -55,7 +68,10 @@ public class SesameContainerService extends AbstractSesameLDPService implements 
 					}
 					break;
 				case MEMBER_RESOURCES:
-					throw new NotImplementedException();
+					if ( containerRetrievalPreferences.contains( APIPreferences.ContainerRetrievalPreference.CONTAINED_RESOURCES ) ) break;
+					Set<IRI> members = containerRepository.getMemberIRIs( containerIRI, orderByRetrievalPreferences );
+					container = getResources( members, container );
+					break;
 				case NON_READABLE_MEMBERSHIP_RESOURCE_TRIPLES:
 					if ( ! containerRetrievalPreferences.contains( APIPreferences.ContainerRetrievalPreference.MEMBERSHIP_TRIPLES ) ) {
 						Set<Statement> membershipTriples = servicesInvoker.proxy( ( proxy ) -> {
@@ -78,6 +94,10 @@ public class SesameContainerService extends AbstractSesameLDPService implements 
 
 	public Set<Statement> getReadableMembershipResourcesTriples( IRI containerIRI ) {
 		return containerRepository.getMembershipTriples( containerIRI );
+	}
+
+	public Set<Statement> getReadableContainedResourcesTriples( IRI containerIRI ) {
+		return containerRepository.getContainmentTriples( containerIRI );
 	}
 
 	public Set<Statement> getNonReadableMembershipResourcesTriples( IRI containerIRI ) {
@@ -148,12 +168,9 @@ public class SesameContainerService extends AbstractSesameLDPService implements 
 
 	@Override
 	public void removeMembers( IRI containerIRI, Set<IRI> members ) {
-		DateTime modifiedTime = DateTime.now();
-		IRI membershipResource = containerRepository.getTypedRepository( this.getContainerType( containerIRI ) ).getMembershipResource( containerIRI );
 		for ( IRI member : members ) {
 			removeMember( containerIRI, member );
 		}
-		sourceRepository.touch( membershipResource, modifiedTime );
 
 	}
 
@@ -161,12 +178,14 @@ public class SesameContainerService extends AbstractSesameLDPService implements 
 	public void removeMember( IRI containerIRI, IRI member ) {
 		deleteInvertedRelation( containerIRI, member );
 		containerRepository.removeMember( containerIRI, member );
+		IRI membershipResource = containerRepository.getTypedRepository( this.getContainerType( containerIRI ) ).getMembershipResource( containerIRI );
+		sourceRepository.touch( membershipResource );
 	}
 
 	@Override
 	public void removeMembers( IRI containerIRI ) {
-		// TODO: Should the resource be touched here?
-		containerRepository.removeMembers( containerIRI );
+		Set<IRI> members = containerRepository.getMemberIRIs( containerIRI );
+		removeMembers( containerIRI, members );
 	}
 
 	@Override
@@ -177,6 +196,30 @@ public class SesameContainerService extends AbstractSesameLDPService implements 
 		}
 	}
 
+	private Container getResources( Set<IRI> sourcesIRIs, Container container ) {
+		Set<RDFSource> sources = sourceService.get( sourcesIRIs );
+		if ( sources == null || sources.isEmpty() ) return container;
+		RDFSource memberSource = sources.iterator().next();
+		container.getBaseModel().addAll( memberSource.getBaseModel() );
+
+		RDFBlankNode responseDescription = getResponseDescription( container );
+
+		for ( RDFSource source : sources ) {
+			ResponseMetaDataFactory.getInstance().create( container, responseDescription, source );
+		}
+		return container;
+	}
+
+	private RDFBlankNode getResponseDescription( Container container ) {
+		ValueFactory valueFactory = SimpleValueFactory.getInstance();
+
+		BNode bNode = valueFactory.createBNode();
+		RDFBlankNode responseDescription = new RDFBlankNode( container.getDocument(), bNode, null );
+		responseDescription.add( RDFSourceDescription.Property.TYPE.getIRI(), ResponseDescriptionDescription.Resource.CLASS.getIRI() );
+		responseDescription.add( RDFSourceDescription.Property.TYPE.getIRI(), RDFResourceDescription.Resource.VOLATILE.getIRI() );
+		return responseDescription;
+	}
+
 	@Override
 	public void delete( IRI targetIRI ) {
 		sourceRepository.delete( targetIRI, true );
@@ -185,7 +228,9 @@ public class SesameContainerService extends AbstractSesameLDPService implements 
 	private void deleteInvertedRelation( IRI containerIRI, IRI memberIRI ) {
 		IRI isMemberOfRelation = resourceRepository.getIRI( containerIRI, ContainerDescription.Property.MEMBER_OF_RELATION );
 		if ( isMemberOfRelation == null ) return;
-		resourceRepository.remove( memberIRI, isMemberOfRelation, containerIRI );
+		IRI membershipResource = containerRepository.getTypedRepository( getContainerType( containerIRI ) ).getMembershipResource( containerIRI );
+		resourceRepository.remove( memberIRI, isMemberOfRelation, membershipResource, memberIRI );
+		sourceRepository.touch( memberIRI );
 	}
 
 	@Autowired